+++ src/components/icons/index.ts
@@ -7,6 +7,7 @@ import RateStarIcon from 'src/assets/svg/mp_act_lt_rate-star.svg?react'
 import RateStarIconDt from 'src/assets/svg/mp_act_dt_rate-star.svg?react'
 import RateUpIcon from 'src/assets/svg/rate_up.svg?react'
 import RateUpIconDt from 'src/assets/svg/mp_act_dt_rate-up.svg?react'
+import RunIcon from 'src/assets/svg/mp_act_lt_run.svg?react'
 import AnalyticsIcon from 'src/assets/svg/mp_pm_lt_analytcis.svg?react'
 import AnalyticsIconDt from 'src/assets/svg/mp_pm_dt_analytics.svg?react'
 import C360Icon from 'src/assets/svg/mp_pm_lt_c360.svg?react'
@@ -40,11 +41,11 @@ import GridIcon from 'src/assets/svg/grid.svg?react'
 import GroupIcon from 'src/assets/svg/mp_info_lt_group.svg?react'
 import HeartIcon from 'src/assets/svg/heart.svg?react'
 import HelpIcon from 'src/assets/svg/help.svg?react'
+import HistoryIcon from 'src/assets/svg/mp_info_lt_history.svg?react'
 import IdentityIcon from 'src/assets/svg/mp_pm_lt_identity.svg?react'
 import JumpToIcon from 'src/assets/svg/mp_act_lt_jump-to.svg?react'
 import LightBulbIcon from 'src/assets/svg/lightbulb.svg?react'
 import LiveStreamIcon from 'src/assets/svg/mp_pm_lt_live-stream.svg?react'
-import PaywallIcon from 'src/assets/svg/paywall.svg?react'
 import MessageQuestionIcon from 'src/assets/svg/message-question.svg?react'
 import MpLogoIcon from 'src/assets/svg/mpLogo.svg?react'
 import ObservabilityIcon from 'src/assets/svg/mp_pm_lt_observability.svg?react'
@@ -63,17 +64,22 @@ import SystemAlertsIcon from 'src/assets/svg/mp_pm_lt_system-alerts.svg?react'
 import TransformationsIcon from 'src/assets/svg/mp_pm_lt_transformations.svg?react'
 import UploadIcon from 'src/assets/svg/mp_act_it_upload.svg?react'
 import UserAttributeIcon from 'src/assets/svg/user-attribute.svg?react'
+import UserIcon from 'src/assets/svg/mp_info_lt_user.svg?react'
 import UserProfilesIcon from 'src/assets/svg/mp_pm_lt_userprofiles.svg?react'
+import ViewOnlyIcon from 'src/assets/svg/mp_info_lt_view-only.svg?react'
 import UsersIcon from 'src/assets/svg/mp_info_lt_users.svg?react'
 import WrenchIcon from 'src/assets/svg/wrench.svg?react'
 import ZoomIn from 'src/assets/svg/mp_act_lt_zoom-in.svg?react'
 import ZoomOut from 'src/assets/svg/mp_act_lt_zoom-out.svg?react'
 import PremiumIcon from 'src/assets/svg/mp_info_lt_premium.svg?react'
+import PaywallIcon from 'src/assets/svg/mp_info_lt_paywall.svg?react'
 import ConversionIcon from 'src/assets/svg/mp_info_lt_conversion.svg?react'
 import HelpVideoIcon from 'src/assets/svg/mp_info_lt_help-video.svg?react'
 import NextIcon from 'src/assets/svg/mp_act_lt_next.svg?react'
 import NumberIcon from 'src/assets/svg/mp_data_lt_number.svg?react'
 import OpenTabIcon from 'src/assets/svg/mp_act_lt_open-tab.svg?react'
+import PauseIcon from 'src/assets/svg/mp_act_lt_pause.svg?react'
+import PlayIcon from 'src/assets/svg/mp_act_lt_play.svg?react'
 import PreviousIcon from 'src/assets/svg/mp_act_lt_previous.svg?react'
 import DirectoryIcon from 'src/assets/svg/mp_pm_lt_directory.svg?react'
 import LockIcon from 'src/assets/svg/mp_act_lt_lock.svg?react'
@@ -129,6 +135,8 @@ import OtherDataIcon from 'src/assets/svg/mp_data_dt_other.svg?react'
 import RefreshIcon from 'src/assets/svg/refresh.svg?react'
 import StringIcon from 'src/assets/svg/mp_data_dt_string.svg?react'
 import TimestampIcon from 'src/assets/svg/mp_data_dt_timestamp.svg?react'
+import ToggleLogicBlocks from 'src/assets/svg/mp_act_lt_toggle-logic-blocks.svg?react'
+import MoveToTop from 'src/assets/svg/mp_act_lt_move-to-top.svg?react'
 
 export {
   AbSplit,
@@ -146,6 +154,7 @@ export {
   RateStarIconDt,
   RateUpIcon,
   RateUpIconDt,
+  RunIcon,
   AnalyticsIcon,
   AnalyticsIconDt,
   Annotation,
@@ -195,6 +204,7 @@ export {
   HeartIcon,
   HelpIcon,
   HelpVideoIcon,
+  HistoryIcon,
   IdentityIcon,
   InfoIcon,
   JourneyAnalysis,
@@ -218,8 +228,9 @@ export {
   OversightIconDt,
   OverviewIcon,
   OverviewIconDt,
-  PaywallIcon,
+  PauseIcon,
   Placeholder,
+  PlayIcon,
   PrecisionIcon,
   PredictionsIcon,
   PredictionsIconDt,
@@ -244,12 +255,15 @@ export {
   UnlockIcon,
   UploadIcon,
   UserAttributeIcon,
+  UserIcon,
   UserProfilesIcon,
   UsersIcon,
+  ViewOnlyIcon,
   WrenchIcon,
   ZoomIn,
   ZoomOut,
   BannerFreemium,
+  PaywallIcon,
   StateEmpty,
   StateNoResults,
   Back,
@@ -262,4 +276,6 @@ export {
   RefreshIcon,
   StringIcon,
   TimestampIcon,
+  ToggleLogicBlocks,
+  MoveToTop,
 }